added components for onboarding screen
